AddressSanitizer support
This catches memory corruption bugs which are normally hard to debug. It slows down execution speed significantly (google estimates 2x slowdown) so shouldn't be enabled unless investigating a bug.

asan complains about RTLD_DEEPBIND so that has to go. Also fixed one usage of Q_min which won't compile when REHLDS_SSE is disabled. math_sse.cpp funcs are causing stack corruption so I also disabled that here.

+++ rehlds/engine/sv_main.cpp
@@ -5647,7 +5647,7 @@ void SV_CreateBaseline(void)
 	MSG_WriteByte(&g_psv.signon, svc_spawnbaseline);
 	MSG_StartBitWriting(&g_psv.signon);
 
-	int sendEdicts = Q_min(sv_max_client_edicts.value, g_psv.num_edicts);
+	int sendEdicts = Q_min((int)sv_max_client_edicts.value, g_psv.num_edicts);
 
 	for (entnum = 0; entnum < sendEdicts; entnum++)
 	{
